Move Element component under Tabs/; Add test for deepRef
