+++ src/components/Tabs/components/Element/index.js
